De-bloat Character, part 6 of X: Move ammo, reloading, and proficiency functions to new cpp files (#51352)
* Move proficiency stuff to character_proficiency.cpp

Including:

- Character::has_proficiency
- Character::get_proficiency_practice
- Character::has_prof_prereqs
- Character::add_proficiency
- Character::lose_proficiency
- Character::display_proficiencies
- Character::practice_proficiency
- Character::proficiency_training_needed
- Character::known_proficiencies
- Character::learning_proficiencies
- Character::get_proficiency_bonus
- Character::set_proficiency_practice

* Move ammo/reload/unload stuff to character_ammo.cpp

Includes:

- Character::find_reloadables
- Character::ammo_count_for
- Character::can_reload
- Character::rate_action_reload
- Character::rate_action_unload
- Character::item_reload_cost
- Character::list_ammo
- Character::select_ammo(item, vector)
- Character::select_ammo(item, bool, bool)
